[refactor] TargetMetadata 클래스 생성
