+++ src/main/kotlin/com/meokq/api/core/enums/TargetType.kt
@@ -1,9 +1,10 @@
-package com.meokq.api.emoji.enums
+package com.meokq.api.core.enums
 
 import com.meokq.api.core.exception.NotFoundException
 
 enum class TargetType {
     QUEST,CHALLENGE,EMOJI;
+
     companion object {
         fun fromString(value: String): TargetType {
             return when (value.uppercase()) {
