+++ src/main/kotlin/com/meokq/api/core/model/TargetMetadata.kt
@@ -1,0 +1,5 @@
+package com.meokq.api.core.model
+
+import com.meokq.api.core.enums.TargetType
+
+class TargetMetadata(val targetType: TargetType, val targetId: String, val userId: String)
