+++ src/main/kotlin/com/meokq/api/emoji/model/Emoji.kt
@@ -2,7 +2,8 @@ package com.meokq.api.emoji.model
 
 import com.meokq.api.core.model.BaseDateTimeModel
 import com.meokq.api.emoji.enums.EmojiStatus
-import com.meokq.api.emoji.enums.TargetType
+import com.meokq.api.core.enums.TargetType
+import com.meokq.api.core.model.TargetMetadata
 import com.meokq.api.emoji.request.EmojiRegisterReq
 import jakarta.persistence.*
 import org.hibernate.annotations.UuidGenerator
@@ -20,31 +21,20 @@ data class Emoji(
     var emojiId : String? = null,
 
     @Enumerated(EnumType.STRING)
-    var status : EmojiStatus? = null,
+    val status : EmojiStatus,
 
-    var userId : String? = null,
+    @Embedded
+    val targetMetadata: TargetMetadata
 
-    @Enumerated(EnumType.STRING)
-    var targetType : TargetType? = null,
+): BaseDateTimeModel() {
 
-    var targetId : String? = null
-): BaseDateTimeModel()
-{
-    fun like(req:EmojiRegisterReq, reqUserId: String) = Emoji (
-        targetId = req.targetId,
-        targetType = convertStatusNameToEnum(req.targetType),
-        userId = reqUserId,
-        status = EmojiStatus.LIKE
+    constructor(status: EmojiStatus, req: EmojiRegisterReq, reqUserId: String ) : this(
+        emojiId = null,
+        status = status,
+        targetMetadata =  TargetMetadata(
+            targetId = req.targetId,
+            targetType = TargetType.fromString(req.targetType),
+            userId = reqUserId)
+
     )
-    fun hate(req:EmojiRegisterReq, reqUserId: String) = Emoji (
-        targetId = req.targetId,
-        targetType = convertStatusNameToEnum(req.targetType),
-        userId = reqUserId,
-        status = EmojiStatus.HATE
-    )
-
-
-    private fun convertStatusNameToEnum(targetName:String): TargetType {
-        return TargetType.fromString(targetName)
-    }
 }
