+++ src/main/kotlin/com/meokq/api/user/request/CustomerXpReq.kt
@@ -1,6 +1,6 @@
 package com.meokq.api.user.request
 
-import com.meokq.api.emoji.enums.TargetType
+import com.meokq.api.core.enums.TargetType
 import com.meokq.api.logs.processor.UserAction
 import org.jetbrains.annotations.NotNull
 
@@ -13,4 +13,6 @@ data class CustomerXpReq(
         xpPoint = userAction.xpPoint,
         title = userAction.title
     )
+
+
 }
